+++ src/Whipstaff.AspNetCore/Whipstaff.AspNetCore.csproj
@@ -44,6 +44,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Whipstaff.Core\Whipstaff.Core.csproj" />
+    <ProjectReference Include="..\Whipstaff.MediatR\Whipstaff.MediatR.csproj" />
     <ProjectReference Include="..\Whipstaff.Runtime\Whipstaff.Runtime.csproj" />
   </ItemGroup>
