+++ src/Whipstaff.MediatR.EntityFrameworkCore/FetchFromEntityFrameworkByLongIdQueryHandler.cs
@@ -9,7 +9,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Whipstaff.Core.Entities;
-using Whipstaff.Core.Mediatr;
 
 namespace Whipstaff.MediatR.EntityFrameworkCore
 {
