+++ src/Whipstaff.MediatR.EntityFrameworkCore/FetchFromEntityFrameworkQueryHandler{TQuery,TDbContext,TEntity,TResult}.cs
@@ -8,7 +8,6 @@ using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using Whipstaff.Core.Mediatr;
 
 namespace Whipstaff.MediatR.EntityFrameworkCore
 {
