+++ src/Whipstaff.MediatR.EntityFrameworkCore/FuncFetchFromEntityFrameworkByLongIdQueryHandler{TQuery,TDbContext,TEntity,TResult}.cs
@@ -8,7 +8,6 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Whipstaff.Core.Entities;
-using Whipstaff.Core.Mediatr;
 
 namespace Whipstaff.MediatR.EntityFrameworkCore
 {
