+++ src/Whipstaff.MediatR.EntityFrameworkCore/InsertUnkeyedEntityIntoKeyedDbSetCommandHandler.cs
@@ -6,7 +6,6 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using Whipstaff.Core.Mediatr;
 
 namespace Whipstaff.MediatR.EntityFrameworkCore
 {
