+++ src/Whipstaff.MediatR.EntityFrameworkCore/Whipstaff.MediatR.EntityFrameworkCore.csproj
@@ -12,6 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Whipstaff.Core\Whipstaff.Core.csproj" />
+    <ProjectReference Include="..\Whipstaff.MediatR\Whipstaff.MediatR.csproj" />
   </ItemGroup>
 
 </Project>
