+++ src/Whipstaff.MediatR/ICommand.cs
@@ -4,7 +4,7 @@
 
 using MediatR;
 
-namespace Whipstaff.Core.Mediatr
+namespace Whipstaff.MediatR
 {
     /// <summary>
     /// Represents a MediatR Command.
