+++ src/Whipstaff.MediatR/RequestPreProcessorRegistrationHandler.cs
@@ -3,24 +3,22 @@
 // See the LICENSE file in the project root for full license information.
 
 using System;
-using MediatR;
 using MediatR.Pipeline;
 
-namespace Whipstaff.Core.Mediatr
+namespace Whipstaff.MediatR
 {
     /// <summary>
-    /// Registers a concrete type for MediatR post processors.
+    /// Registers a concrete type for MediatR pre processors.
     /// </summary>
     /// <typeparam name="TImplementationType">The type for the request handler.</typeparam>
     /// <typeparam name="TRequest">The type for the mediatr request.</typeparam>
-    /// <typeparam name="TResponse">The type for the mediatr response.</typeparam>
-    public sealed class RequestPostProcessorRegistrationHandler<TImplementationType, TRequest, TResponse>
-        : IRequestPostProcessorRegistrationHandler
-        where TImplementationType : class, IRequestPostProcessor<TRequest, TResponse>
-        where TRequest : IRequest<TResponse>
+    public sealed class RequestPreProcessorRegistrationHandler<TImplementationType, TRequest>
+        : IRequestPreProcessorRegistrationHandler
+        where TImplementationType : class, IRequestPreProcessor<TRequest>
+        where TRequest : notnull
     {
         /// <inheritdoc/>
-        public Type ServiceType => typeof(IRequestPostProcessor<TRequest, TResponse>);
+        public Type ServiceType => typeof(IRequestPreProcessor<TRequest>);
 
         /// <inheritdoc/>
         public Type ImplementationType => typeof(TImplementationType);
