+++ src/Whipstaff.Testing/Cqrs/FakeAuditableCommandFactory.cs
@@ -6,6 +6,7 @@ using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using Whipstaff.Core;
+using Whipstaff.MediatR;
 
 namespace Whipstaff.Testing.Cqrs
 {
