+++ src/Whipstaff.Testing/Cqrs/FakeAuditableQueryFactory.cs
@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 using Whipstaff.Core;
+using Whipstaff.MediatR;
 
 namespace Whipstaff.Testing.Cqrs
 {
