+++ src/Whipstaff.UnitTests/Features/Mediatr/MediatrHelperTests.cs
@@ -13,9 +13,9 @@ using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Whipstaff.Core.Mediatr;
 using Whipstaff.EntityFramework.ModelCreation;
 using Whipstaff.EntityFramework.RowVersionSaving;
+using Whipstaff.MediatR;
 using Whipstaff.Testing.Cqrs;
 using Whipstaff.Testing.EntityFramework;
 using Whipstaff.Testing.MediatR;
